For merging

+++ swiftmail_script.py
@@ -8,30 +8,37 @@ import busio
 import adafruit_vcnl4010
 import adafruit_tsl2591
 
-
+# --------------------
 # SENSOR THRESHOLDS
+# --------------------
 LUX_OPEN_THRESHOLD = 18
 PROX_NEW_MAIL_THRESHOLD = 8000
 
-
+# --------------------
 # MQTT CONFIG
+# --------------------
 broker = "test.mosquitto.org"
 BASE_TOPIC = "swiftmail"
 
-
+# --------------------
 # I2C + SENSORS
+# --------------------
 i2c = busio.I2C(board.SCL, board.SDA)
 lux_sensor = adafruit_tsl2591.TSL2591(i2c)
 prox_sensor = adafruit_vcnl4010.VCNL4010(i2c)
 
+# --------------------
 # MQTT CALLBACKS
+# --------------------
 def on_connect(client, userdata, flags, rc):
     print("Connected with result code:", rc)
 
 def on_disconnect(client, userdata, rc):
     print("Disconnected with result code:", rc)
 
+# --------------------
 # MQTT CLIENT
+# --------------------
 client = mqtt.Client()
 client.on_connect = on_connect
 client.on_disconnect = on_disconnect
@@ -40,20 +47,24 @@ print("Connecting to broker:", broker)
 client.connect(broker, 1883, 60)
 client.loop_start()
 
-# INITIAL STATE MEMORY
+# --------------------
+# STATE MEMORY
+# --------------------
 last_door_state = None
 last_mail_state = None
 
+# --------------------
 # MAIN LOOP
+# --------------------
 while True:
-    lux = round(Decimal(lux_sensor.lux), 2)
+    lux = float(round(Decimal(lux_sensor.lux), 2))
     proximity = prox_sensor.proximity
 
     # Publish raw sensor values
     client.publish(f"{BASE_TOPIC}/raw/lux", lux, retain=True)
     client.publish(f"{BASE_TOPIC}/raw/proximity", proximity, retain=True)
 
-    # Detect door events
+    # Detect door state
     door_state = "OPEN" if lux > LUX_OPEN_THRESHOLD else "CLOSED"
 
     if door_state != last_door_state:
@@ -65,7 +76,7 @@ while True:
         )
         last_door_state = door_state
 
-    # Detect mail events
+    # Detect mail state
     mail_state = "NEW_MAIL" if proximity > PROX_NEW_MAIL_THRESHOLD else "NO_MAIL"
 
     if mail_state != last_mail_state:
@@ -77,9 +88,9 @@ while True:
         )
         last_mail_state = mail_state
 
-    # Pinging if the device is alive (heartbeat)
+    # Heartbeat (device alive)
     timestamp = datetime.datetime.now().isoformat()
-    client.publish(
+        client.publish(
         f"{BASE_TOPIC}/heartbeat",
         timestamp,
         retain=True
